requirements.txt
fastapi
httpx
openai
python-dotenv
qdrant-client
requests
sentence-transformers
sentencepiece
tiktoken
uvicorn
